// examples/typescript-server/src/generated/services/vehicle-service.ts
// GENERATED CODE - DO NOT MODIFY BY HAND. ANY MODIFICATION WILL BE LOST ON NEXT GENERATION
// Generated by GraphLink dev
// GitHub: https://github.com/Oualitsen/graphlink
// Site: https://graphlink.dev
// Pub.dev https://pub.dev/packages/graphlink

import { GraphLinkContext } from '../context.js';
import { GraphQLResolveInfo } from 'graphql';
import { Vehicle } from '../types/vehicle.js';
import { AddVehicleInput } from '../inputs/add-vehicle-input.js';

export interface VehicleService {
  getVehicle(id: string, context: GraphLinkContext, info: GraphQLResolveInfo): Promise<Vehicle | null>;
  listVehicles(context: GraphLinkContext, info: GraphQLResolveInfo): Promise<Vehicle[]>;
  addVehicle(input: AddVehicleInput, context: GraphLinkContext, info: GraphQLResolveInfo): Promise<Vehicle>;
  vehicleStatusChanged(vehicleId: string, context: GraphLinkContext): AsyncIterable<Vehicle>;
}