// examples/typescript-server/src/generated/types/vehicle.ts
// GENERATED CODE - DO NOT MODIFY BY HAND. ANY MODIFICATION WILL BE LOST ON NEXT GENERATION
// Generated by GraphLink dev
// GitHub: https://github.com/Oualitsen/graphlink
// Site: https://graphlink.dev
// Pub.dev https://pub.dev/packages/graphlink

import { VehicleStatus } from '../enums/vehicle-status.js';


export interface Vehicle {
   readonly id: string | null;
   readonly brand: string | null;
   readonly year: number | null;
   readonly status: VehicleStatus | null;
}