// examples/typescript-server/src/generated/inputs/add-vehicle-input.ts
// GENERATED CODE - DO NOT MODIFY BY HAND. ANY MODIFICATION WILL BE LOST ON NEXT GENERATION
// Generated by GraphLink dev
// GitHub: https://github.com/Oualitsen/graphlink
// Site: https://graphlink.dev
// Pub.dev https://pub.dev/packages/graphlink



export interface AddVehicleInput {
   brand: string;
   year: number;
   ownerId: string;
}