// examples/typescript-server/src/generated/enums/vehicle-status.ts
// GENERATED CODE - DO NOT MODIFY BY HAND. ANY MODIFICATION WILL BE LOST ON NEXT GENERATION
// Generated by GraphLink dev
// GitHub: https://github.com/Oualitsen/graphlink
// Site: https://graphlink.dev
// Pub.dev https://pub.dev/packages/graphlink



export enum VehicleStatus {
   ACTIVE = 'ACTIVE',
   INACTIVE = 'INACTIVE',
   SOLD = 'SOLD',
}